GoodNight
switched over to vs code. I want pluralsight to record my progress and vs code is the only way to do that. Visual studio does not have that option. Although Visual Studio is a much better option, atleast for C# ii is. but i need the exp points on pluralsight.... also got stuck on some shit with the debugger and the terminal.